Redirect HTTP to HTTPS if available. #RN
This method is not 100% clean. Consider this:
An Apache server HAS mod_ssl activated but the actual site isn't
set up to run via HTTPS. So either disable HTTPS completely or
set up the site for HTTPS.
A more clean way would be to activate the redirection only if the
site has HTTPS as well.